Changes after round 3 testing

+++ Actions/Delete.py
@@ -1,12 +1,14 @@
 import os
+from Actions import Search
 from google.cloud import storage
 
 
 def delete_object_safe(bucket_name, object_name):
-    storage_client = storage.Client()
-    bucket = storage_client.bucket(bucket_name)
-    blob = bucket.blob(object_name)
-    blob.delete()
+    if Search.find_object(bucket_name, object_name):
+        storage_client = storage.Client()
+        bucket = storage_client.bucket(bucket_name)
+        blob = bucket.blob(object_name)
+        blob.delete()
 
 
 def delete_object(bucket_name, object_name):
